fix(api): move cookie type inteface to root interface folder
