+++ packages/api/src/modules/auth/auth.controller.ts
@@ -9,10 +9,10 @@ import {
   UnauthorizedException,
 } from '@nestjs/common';
 
+import { CookieRequest, CookieResponse } from '../../interfaces/cookie-types';
 import { BasePayload, TokenPayload } from '../../interfaces/token-types';
 import { ValidationPipe } from '../../utils/validation.pipe';
 import { AuthService } from './auth.service';
-import { CookieRequest, CookieResponse } from './cookie/interfaces/cookie-types';
 import { Auth } from './decorators/auth.decorator';
 import { CookiePayload } from './decorators/cookiePayload.decorator';
 import { LoginDto } from './dto/login.dto';
